refactor : notification도메인 dto 일부 리펙토링
