+++ src/main/java/CamNecT/server/global/notification/dto/response/RegisterPushTokenResponse.java
@@ -1,4 +1,4 @@
-package CamNecT.server.global.notification.dto;
+package CamNecT.server.global.notification.dto.response;
 
 public record RegisterPushTokenResponse(
         Long pushDeviceId,
